requirements.txt
google-generativeai
halo
openai>=1.0.0